refactor(block): extracted code to submodules
